requirements.txt
Flask
flask-cors
requests
python-dotenv
urllib3